[exp] Introduced a seam for a coordinator.

+++ exp/run.py
@@ -14,11 +14,14 @@ from scripts.network import Network
 
 def run(args: Namespace) -> None:
     path = Path(args.path)
+    name = args.name
 
-    shutil.rmtree(path, ignore_errors=True)
-    os.system(f"mkdir -p {path}")
+    if not args.leader_port:  
+        shutil.rmtree(path, ignore_errors=True)
+        os.system(f"mkdir -p {path}")
 
     controller = Controller(
+        name = name,
         network = Network(
             bandwidth=getattr(args, 'bandwidth', None),
             delay=getattr(args, 'delay', None),
@@ -28,31 +31,42 @@ def run(args: Namespace) -> None:
         only_server = args.only_server,
         port = args.port,
         path = path,
+        leader_port = args.leader_port,
     )
-    plots = {
-        'qoe' : LivePlot(figure_name='qoe', y_label='qoe'),
-        'rebuffer' : LivePlot(figure_name='rebuffer', y_label='rebuffer'),
-        'switch' : LivePlot(figure_name='switch', y_label='switch'),
-        'quality' : LivePlot(figure_name='quality', y_label='quality'),
-    }
-    (Server('experiment', args.port)
+    
+    server = Server('experiment', args.port)
+        
+    request_port = args.leader_port if args.leader_port else args.port 
+    (server
         .add_post('/init', controller.on_init())
         .add_post('/start', controller.on_start())
-        .add_post('/metrics', Monitor(path))
-        .add_post('/qoe', plots['qoe'])
-        .add_post('/rebuffer', plots['rebuffer'])
-        .add_post('/switch', plots['switch'])
-        .add_post('/quality', plots['quality'])
-        .add_post('/complete', multiple_sync(
-            OnComplete(path, plots), 
-            controller.on_complete(),
-        )) 
-        .add_post('/destroy', controller.on_destroy())
-        .run())
+        .add_post('/metrics', Monitor(path, args.plot, request_port))
+        .add_post('/destroy', controller.on_destroy()))
+
+    plots = {}
+    if args.plot:
+        plots = {
+            'qoe' : LivePlot(figure_name='qoe', y_label='qoe'),
+            'rebuffer' : LivePlot(figure_name='rebuffer', y_label='rebuffer'),
+            'switch' : LivePlot(figure_name='switch', y_label='switch'),
+            'quality' : LivePlot(figure_name='quality', y_label='quality'),
+        }
+        (server
+            .add_post('/qoe', plots['qoe'])
+            .add_post('/rebuffer', plots['rebuffer'])
+            .add_post('/switch', plots['switch'])
+            .add_post('/quality', plots['quality']))
+
+    server.add_post('/complete', multiple_sync(
+        OnComplete(path, name, plots), 
+        controller.on_complete(),
+    )) 
+    server.run()
 
 
 if __name__ == "__main__":
     parser = ArgumentParser(description='Run an experiment.')
+    parser.add_argument('--name', type=str, default='default', help='Instance name. Must be specified when running with a leader.')
     parser.add_argument('--port', type=int, default=8080, help='Port(default 8080).')
     parser.add_argument('--path', type=str, default='logs/default', help='Experiment folder path.')
     parser.add_argument('--only-server', dest='only_server', action='store_true', help='Use only as metrics server.')
@@ -60,4 +74,6 @@ if __name__ == "__main__":
     parser.add_argument('-b', '--bandwidth', type=float, help='Bandwidth of the link.')
     parser.add_argument('-t', '--trace', type=str, help='Trace of bandwidth.')
     parser.add_argument('-d','--dash', action='append', help='Add arguments to dash player.')
+    parser.add_argument('--plot', action='store_true', help='Enable plotting.')
+    parser.add_argument('-lp', '--leader-port', dest='leader_port', type=int, required=False, help='Port of the leader.')
     run(parser.parse_args())
